Implement games table scan feature/tests, remove GameControllerTests
GameRepository now has methods to scan and retrieve the entire games
table. Please note that there is no result limit in place, so this may
be important to add in the future depending on how large the table
becomes.

After due deliberation and research, I decided to remove
GamesControllerTests entirely, as There is little purpose in writing
unit tests for a class whose methods are simply wrappers around the
already well-tested GamesRepository class. If the complexity of
GameController increases in the future or if conditional status
messages were to be implemented, unit tests may become warranted.

As an alternative, Controller methods/endpoints and repository
integration should be validated using feature-level and integration
tests. I intend for this to be done sooner, rather than later.