Implement combat system and Entity classes, rename Game to Battle
The Battle simulation code has been implemented such that it is
"just good enough" to allow the Player to engage in battle. However,
even in its infancy, this code is begging to be refactored due to
the amount of code duplication and nesting.

Assuming this is not simply by nature of the combat system, it is
likely that an implementation based on a queue or a n^2 lookup table
would be a viable alternative.

As an aside, I am aware of the sizes of these recent commits. In the
early stages of this project, it is difficult to make feature additions
without incurring large swaths of changes throughout the codebase.
Perhaps this points to a larger issue regarding coupling, or perhaps
there is a better way to make incremential changes even on these large
features.