Add tests and fixes for dodge logic
BattleManager attemptDodge() function replaced with public coinFlip()
method, allowing Bernoulli trials to be performed from anywhere and with
custom odds. Easier to test and more useful.

While fixing a failing dodge test, it was discovered that test outcomes
can sometimes be affected by random move order, which itself revealed
a section of improperly nested code. The implications are two-fold:

First, either simulateBattle needs to be rewritten such that turn order
does not affect which sections of code are executed, or move order needs
to be accounted for with multiple versions of nearly every test.

Second, new feature development needs to be moved onto separate
branches so that any fixes which themselves reveal more issues can be
pushed without resulting in "yak shaving" to maintain the stability of
the main branch.