Rewrite BattleManager
Note: MANY TESTS ARE CURRENTLY FAILING. Be cautious when resetting to
this commit!!

This is a behemoth of a commit. My apologies. Here's a list of changes:
* Changed BattleManager to a static class
* Implemented actionHandler Map
    - takes a Set of two BattleActions and returns a class methojd
    - designed to be agnostic to ordering of actions
* Implemented BMFunc FunctionalInterface within BattleManager
    - the purpose of this is to allow anonymously calling functions
      returned by the actionHandler Map
* Added playerAction parameter to most action handler methods and
  implemented fixed Entity parameter order
    - allows methods to compute which Entity performed what action,
      which significantly cuts down on the size and complexity of
      actionHandler
* Refactored simulateBattle method to be blind to the specific actions
  performed
    - this significantly improves the issues of readability/complexity
    - fixes a bug where move priority was not resetting between turns
* Changed reportActions to single-action reportAction

When redesigning simulateBattle, I was presented with a philosophical
choice between two routes:
A. Design actionHandler to be agnostic to action order and increase
 complexity of handler methods
B. Allow handler methods to stay the same at the cost of having the Map
 be either twice as large or twice as complex.

I chose A. My reasoning is that actionHandler is large and difficult
enough to read as it is. Adding aroundmore keys that flip the order
of actions adds significantly more lines of code and requires the user
to be much more careful when getting a handler method from it. The
other option would be to rewrite the lambdas to use if-else statements,
which is even more difficult to read.

The downside is that handler methods must now compute which entity is
performing what action. This is facilitated by a new playerAction param
and a check at the beginning to assign roles. I figured this was worth
it. Going to keep my eyes peeled for a more elegant implementation.

Once again, sorry for shoving so much into a single commit.