Rewrite BattleManager as non-static, add Battle interface methods
I decided that it would be better design all-around if I made
BattleManager an instanced class. The main benefits of doing so are
simplified signatures for all handler methods and an easier class to
test, given that Mockito cannot stub static classes.

The only potential issue I foresee here is that BattleController must
create a new BattleManager instance every time it is called. This should
be of little consequence, hopefully.

I also added an IllegalArgumentException to Battle.setPriority to
control which values it can be given. Hoping to replace the integer
system with an "Ally", "Enemy", "Neutral" system in the future.