Replace BattleManager coinFlip() with Entity dodge() method
This change was made to allow for easier testing. At the moment,
Entity dodgeChance is a random number thrown into the bottom half of
the file.

The intent behind this change is to enable easier testing and better
dependency injection.

I also renamed parameters in handleAttackAttack to be more intuitive.