Readd mistakenly removed setBattleID method to Battle
Method was originally removed to prevent Battle hash key values from
ever changing. This caused all BattleRepository tests to fail, but
this was not discovered until now because I assumed my changes in the
last commits were benign. Lesson learned.

Going to look for an implementation that ensures battleID is a final
attribute in the future.